M1/services/data-service/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "civiclens-data-service"
version = "0.1.0"
description = "M1 data-service for CivicLens"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.111,<1.0",
  "uvicorn[standard]>=0.30,<1.0",
  "sqlalchemy>=2.0,<3.0",
  "alembic>=1.13,<2.0",
  "pydantic-settings>=2.2,<3.0",
  "psycopg[binary]>=3.1,<4.0",
  "httpx>=0.27,<1.0",
]

[tool.setuptools]
packages = ["app", "app.routers"]

[tool.ruff]
line-length = 100

[tool.pytest.ini_options]
pythonpath = ["."]